sets up karma-jasmin test environment,
