+++ package.json
@@ -43,10 +43,8 @@
   "devDependencies": {
     "angular-mocks": "^1.5.9",
     "autoprefixer": "^6.0.3",
-    "babel": "^6.5.2",
     "babel-core": "^6.18.2",
     "babel-loader": "^6.2.0",
-    "babel-polyfill": "^6.16.0",
     "babel-preset-es2015": "^6.18.0",
     "copy-webpack-plugin": "^3.0.1",
     "css-loader": "^0.23.0",
@@ -59,20 +57,13 @@
     "karma": "^1.3.0",
     "karma-babel-preprocessor": "^6.0.1",
     "karma-chrome-launcher": "^2.0.0",
-    "karma-coverage": "^1.0.0",
     "karma-jasmine": "^1.0.2",
-    "karma-phantomjs-launcher": "^1.0.0",
-    "karma-requirejs": "^1.1.0",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-spec-reporter": "0.0.26",
     "karma-webpack": "^1.8.0",
     "node-libs-browser": "^1.0.0",
     "null-loader": "^0.1.1",
     "phantomjs-prebuilt": "^2.1.4",
     "postcss-loader": "^0.9.1",
     "raw-loader": "^0.5.1",
-    "require": "^2.4.20",
-    "requirejs": "^2.3.2",
     "rimraf": "^2.5.1",
     "style-loader": "^0.13.0",
     "webpack": "^1.12.13",
