--- src/app/hello.js
@@ -1,10 +1,0 @@
-'use strict';
-export default function hello() {
-    let name = 'Foo';
-    let greeting = `Hello ${name}`;
-    if (false) {
-        // Should not be covered
-        return 'Good bye';
-    }
-    return greeting;
-}
