--- src/app/hello.spec.js
@@ -1,7 +1,0 @@
-'use strict';
-import hello from './hello';
-describe('hello', () => {
-    it('should return Hello Foo', function () {
-        expect(hello()).toEqual('Hello Foo');
-    });
-});
