--- src/app/login/user_service.spec.js
@@ -1,23 +1,0 @@
-describe('user_service', function() {
-
-  it('has a dummy spec to test 2+2', function() {
-    expect(2+2).toEqual(4);
-  });
-
-  it('ensures user can log in', function() {
-    // expect current scope to contain username
-  });
-  it('ensures path has changed', function() {
-    // expect path to equal '/dashboard'
-  });
-  //critical
-  it('ensures invalid username is caught', function() {});
-  it('ensures valid username pass validation', function() {});
-  it('ensures submitting form changes path', function() { });
-  //nice
-  it('ensures client-side helper message shown for empty fields', function() {})
-  it('ensures hitting enter on password field submits for', function() {})
-
-
-
-});
